trying to fix tests
